Tagging Debian release 1.2.1-2. - This release contains patch for compiling against gcc/g++ 4.6.

+++ dmtcp/src/ptracewrapper.cpp
@@ -53,8 +53,7 @@ extern "C" int ptrace_info_list_size() {
 }
 
 extern "C" struct ptrace_info get_next_ptrace_info(int index) {
-  if ((unsigned int)index >= ptrace_info_list.size())
-    return EMPTY_PTRACE_INFO;
+  if (index >= ptrace_info_list.size()) return EMPTY_PTRACE_INFO;
 
   dmtcp::list<struct ptrace_info>::iterator it;
   int local_index = 0;
